Rename betty.console.assertion to betty.argparse
